Add files via upload
In the popular and brilliant science fiction work of Cixin Liu, The Dark Forest, the protagonist must contend with the stark realization that there is limited space in the universe, that extraterrestrial life is abundant, and therefore resource conflict is inevitable. Coupled with a communication breakdown, conceptualized as ‘a chain of suspicion,’ due to differences and distance that render cooperation impossible, the characters find themselves in a social dilemma on a cosmic scale. Scaled down in much less fantastical fashion, humans cope with social dilemmas of varied proportions daily. They apply to a broad range of problems between as little as two agents, small groups, or entire nations (Van Lange et al., 2013). Many of the contentious social dilemmas that humanity deals with involve conflict between self-interest and the good of the collective. This is seen in commercial overfishing, deforestation, pollution, and illegal activity such as drug and human trafficking. As mankind contends with solving moral dilemmas, they are increasingly relying on artificial intelligence as decision support tools. The concept of an artificial moral agent (AMA) is proposed by Wallach & Allen (2009) and requires a study of ethical theory to hypothesize on the limitations of AI to handle moral dilemma. We will explore this idea in-depth in this paper, consider classical moral dilemmas, and modern advancements on classic matrix game social dilemmas that provide insight to the nature of social cooperation.